Plugins: Adjustments after merging BinCode...
* Generate encoding methods for plugin types.
* Include plugin types in protocol test.
* Adjust numbers types on wit and rust size to make it compatible with
  typescript ones.

+++ application/apps/indexer/stypes/src/plugins/mod.rs
@@ -94,8 +94,8 @@ pub struct PluginConfigItem {
 )]
 pub enum PluginConfigValue {
     Boolean(bool),
-    Number(i64),
-    Float(f64),
+    Number(i32),
+    Float(f32),
     Text(String),
     /// A string representing a file or directory path.
     Path(PathBuf),
